Feature/uc12 productcategoriën toevoegen

+++ Grocery.Core/Models/Product.cs
@@ -7,10 +7,10 @@ namespace Grocery.Core.Models
         [ObservableProperty]
         public int stock;
         public DateOnly ShelfLife { get; set; }
-        public float Price { get; set; }
-        public Product(int id, string name, int stock) : this(id, name, stock, default, 0.0f) { }
+        public decimal Price { get; set; }
+        public Product(int id, string name, int stock) : this(id, name, stock, default, 0.0m) { }
 
-        public Product(int id, string name, int stock, DateOnly shelfLife, float price) : base(id, name) 
+        public Product(int id, string name, int stock, DateOnly shelfLife, decimal price) : base(id, name) 
         {
             Stock = stock;
             ShelfLife = shelfLife;
